Adds project structure around existing code
With this in place, you can enable travis, which can test pull requests
in case they break the build.

Here's a list of other things:

* Adds maven build structure
* Defaults language level to 1.8, except main code in core which is 1.7
* Adds apache license and plugin to verify it
* Reformats code to square style (similar to google style)
* Adds an example unit test we can use when there are concrete types
