+++ .travis.yml
@@ -1,0 +1,15 @@
+# Run `travis lint` when changing this file to avoid breaking the build.
+sudo: false
+
+cache:
+  directories:
+  - $HOME/.m2
+
+language: java
+
+jdk:
+  - oraclejdk8
+
+script:
+  # Only build on pull requests for now
+  - '[ "${TRAVIS_PULL_REQUEST}" = "false" ] || ./mvnw install -nsu'
