--- opentracing-core/src/java/io/opentracing/MarshaledPairBinary.java
@@ -1,12 +1,0 @@
-package io.opentracing;
-
-/**
- * MarshaledPairBinary represents the TraceContext id as well as any "trace tags" in an implementation-specific binary format.
- *
- * @see TraceContextMarshaler
- * @see TraceContextUnmarshaler
- */
-public interface MarshaledPairBinary {
-    byte[] getTraceContextId();
-    byte[] getTraceTags();
-}
