--- opentracing-core/src/java/io/opentracing/SomeMicrosecondType.java
@@ -1,4 +1,0 @@
-package io.opentracing;
-
-// XXX: I'm not sure how to fashionably represent microseconds in Java. This is a standin.
-public interface SomeMicrosecondType {}
