--- opentracing-core/src/java/io/opentracing/TraceContextUnmarshaler.java
@@ -1,9 +1,0 @@
-package io.opentracing;
-
-/**
- * @see TraceContextMarshaler
- */
-public interface TraceContextUnmarshaler {
-    TraceContext unmarshalTraceContextBinary(MarshaledPairBinary m);
-    TraceContext unmarshalTraceContextKeyValue(MarshaledPairKeyValue m);
-}
